+++ src/main/java/com/bongsco/api/employee/repository/RankRepository.java
@@ -1,4 +1,4 @@
-package com.bongsco.api.adjust.common.repository;
+package com.bongsco.api.employee.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
